test: replace deno_mock to msw
