+++ projects/archive.test.ts
@@ -1,21 +1,14 @@
 import { archive, unarchive } from "./archive.ts";
-import { MockFetch } from "https://deno.land/x/deno_mock_fetch@1.0.2/mod.ts";
-import { join } from "jsr:@std/path@1.0.2";
-import { assert, assertEquals } from "jsr:@std/assert@0.226.0";
+import { assert } from "jsr:@std/assert@0.225.2";
 
-const context = {
-  apiKey: "sample",
-  endpoint: "http://readmine.example.com",
-};
+import { context, handler } from "./archive.mock.ts";
+import { setupServer } from "npm:msw@2.3.0/node";
+
+const server = setupServer(...handler);
+server.listen();
 
 Deno.test("test for redmine project 'archive' endpoint", async (t) => {
-  const mockFetch = new MockFetch();
   await t.step("if got 200, should be success", async () => {
-    mockFetch
-      .intercept(join(context.endpoint, "projects", `${1}`, "archive.json"), {
-        method: "PUT",
-      })
-      .response(JSON.stringify({}), { status: 200 });
     const e = await archive(1, context);
     assert(e.isOk());
   });
@@ -23,51 +16,19 @@ Deno.test("test for redmine project 'archive' endpoint", async (t) => {
   await t.step(
     "if get invalid response with error object, should be err with error text",
     async () => {
-      const errorSample = { errors: ["sample error"] };
-      mockFetch
-        .intercept(join(context.endpoint, "projects", `${1}`, "archive.json"), {
-          method: "PUT",
-        })
-        .response(JSON.stringify(errorSample), {
-          status: 422,
-          statusText: "Unprocessable Entity",
-        });
-      const e = await archive(
-        1,
-        context,
-      );
+      const e = await archive(2, context);
       assert(e.isErr());
-      assertEquals(
-        e.error.message,
-        JSON.stringify(errorSample.errors),
-      );
     },
   );
 
   await t.step("if get invalid response with unexpected format", async () => {
-    mockFetch
-      .intercept(join(context.endpoint, "projects", `${1}`, "archive.json"), {
-        method: "PUT",
-      })
-      .response(JSON.stringify("unknown error"), {
-        status: 404,
-        statusText: "Not found",
-      });
-    const e = await archive(1, context);
+    const e = await archive(3, context);
     assert(e.isErr());
   });
-
-  mockFetch.deactivate();
 });
 
 Deno.test("test for redmine project 'unarchive' endpoint", async (t) => {
-  const mockFetch = new MockFetch();
   await t.step("if got 200, should be success", async () => {
-    mockFetch
-      .intercept(join(context.endpoint, "projects", `${1}`, "unarchive.json"), {
-        method: "PUT",
-      })
-      .response(JSON.stringify({}), { status: 200 });
     const e = await unarchive(1, context);
     assert(e.isOk());
   });
@@ -75,45 +36,13 @@ Deno.test("test for redmine project 'unarchive' endpoint", async (t) => {
   await t.step(
     "if get invalid response with error object, should be err with error text",
     async () => {
-      const errorSample = { errors: ["sample error"] };
-      mockFetch
-        .intercept(
-          join(context.endpoint, "projects", `${1}`, "unarchive.json"),
-          {
-            method: "PUT",
-          },
-        )
-        .response(JSON.stringify(errorSample), {
-          status: 422,
-          statusText: "Unprocessable Entity",
-        });
-      const e = await unarchive(
-        1,
-        context,
-      );
+      const e = await unarchive(2, context);
       assert(e.isErr());
-      assertEquals(
-        e.error.message,
-        JSON.stringify(errorSample.errors),
-      );
     },
   );
 
   await t.step("if get invalid response with unexpected format", async () => {
-    mockFetch
-      .intercept(join(context.endpoint, "projects", `${1}`, "unarchive.json"), {
-        method: "PUT",
-      })
-      .response(JSON.stringify("unknown error"), {
-        status: 404,
-        statusText: "Not found",
-      });
-    const e = await unarchive(
-      1,
-      context,
-    );
+    const e = await unarchive(3, context);
     assert(e.isErr());
   });
-
-  mockFetch.deactivate();
 });
