+++ projects/update.test.ts
@@ -1,39 +1,12 @@
 import { update } from "./update.ts";
-import { MockFetch } from "https://deno.land/x/deno_mock_fetch@1.0.2/mod.ts";
-import { join } from "jsr:@std/path@1.0.2";
-import { assert, assertEquals } from "jsr:@std/assert@0.226.0";
+import { context, handler } from "./update.mock.ts";
+import { setupServer } from "npm:msw@2.3.0/node";
+import { assert } from "jsr:@std/assert@0.225.2";
 
-const urlTable = new Map([
-  [1, {
-    status: 200,
-    obj: {},
-  }],
-  [2, {
-    status: 422,
-    obj: { errors: ["some error"] },
-  }],
-  [3, {
-    status: 404,
-    obj: "not found",
-  }],
-]);
-
-const context = {
-  apiKey: "sample",
-  endpoint: "http://readmine.example.com",
-};
+const server = setupServer(...handler);
+server.listen();
 
 Deno.test("test for redmine project 'update' endpoint", async (t) => {
-  const mockFetch = new MockFetch();
-  for (const [id, { status, obj }] of urlTable.entries()) {
-    mockFetch
-      .intercept(join(context.endpoint, "projects", `${id}.json`), {
-        method: "PUT",
-      })
-      .response(JSON.stringify(obj), { status })
-      .persist();
-  }
-
   await t.step("if got 200, should be success", async () => {
     const e = await update(1, { name: "sample" }, context);
     assert(e.isOk());
@@ -44,10 +17,6 @@ Deno.test("test for redmine project 'update' endpoint", async (t) => {
     async () => {
       const e = await update(2, { name: "sample" }, context);
       assert(e.isErr());
-      assertEquals(
-        e.error.message,
-        JSON.stringify((urlTable.get(2)?.obj as { errors: string[] }).errors),
-      );
     },
   );
 
@@ -55,6 +24,4 @@ Deno.test("test for redmine project 'update' endpoint", async (t) => {
     const e = await update(3, { name: "sample" }, context);
     assert(e.isErr());
   });
-
-  mockFetch.deactivate();
 });
